+++ test/blackbox-tests/test-cases/github734/run.t
@@ -1,5 +1,4 @@
   $ jbuilder build @foo
   File "src/dune", line 4, characters 10-17:
-  Error: "stubs/x" is not a valid C name.
-  Hint: To use C files from another directory, use a (copy_files <dir>/*.c) stanza instead.
+  Error: File src/stubs/x.c is not part of the current directory group. This is not allowed.
   [1]